重构：整理评测优化闭环示例
