+++ examples/optimization/eval_optimize_loop/core/__init__.py
@@ -1,0 +1,13 @@
+"""Evaluation and optimization pipeline implementation."""
+
+from .pipeline import prepare_run
+from .pipeline import run_offline_stage
+from .pipeline import run_real_stage
+from .pipeline import run_trace_stage
+
+__all__ = [
+    "prepare_run",
+    "run_offline_stage",
+    "run_real_stage",
+    "run_trace_stage",
+]
