+++ examples/optimization/eval_optimize_loop/data/__init__.py
@@ -1,0 +1,9 @@
+"""Data contracts and bundled inputs for the evaluation optimization example."""
+
+from .config import PipelineConfig
+from .config import load_pipeline_config
+
+__all__ = [
+    "PipelineConfig",
+    "load_pipeline_config",
+]
