--- examples/optimization/eval_optimize_loop/fake/__init__.py
@@ -1,16 +1,0 @@
-# Tencent is pleased to support the open source community by making tRPC-Agent-Python available.
-#
-# Copyright (C) 2026 Tencent. All rights reserved.
-#
-# tRPC-Agent-Python is licensed under the Apache License, Version 2.0.
-"""Deterministic offline components for the eval/optimization loop."""
-
-from .candidate_provider import DeterministicFakeCandidateProvider
-from .model import DeterministicFakeModel
-from .model import deterministic_response
-
-__all__ = [
-    "DeterministicFakeCandidateProvider",
-    "DeterministicFakeModel",
-    "deterministic_response",
-]
