+++ examples/optimization/eval_optimize_loop/sample_output/inputs/pipeline_config.json
@@ -5,9 +5,9 @@
     "candidate_scenario": "improve"
   },
   "inputs": {
-    "train_evalset": "data/train.evalset.json",
-    "validation_evalset": "data/val.evalset.json",
-    "optimizer_config": "optimizer.json"
+    "train_evalset": "data/evalsets/train.evalset.json",
+    "validation_evalset": "data/evalsets/val.evalset.json",
+    "optimizer_config": "configs/optimizer.json"
   },
   "prompts": [
     {
